+++ workers/ephemeral/javascript-plugins-sandbox/src/types/api/v1/api_pb.js
@@ -8864,7 +8864,7 @@ proto.api.v1.Block.prototype.hasSend = function() {
  * @private {!Array<!Array<number>>}
  * @const
  */
-proto.api.v1.Step.oneofGroups_ = [[2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,58,59,60,61,62,63,64,65,66,67,68,69,70,71,72,73,74,75,76,77,78,79,80,81]];
+proto.api.v1.Step.oneofGroups_ = [[2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,58,59,60,61,62,63,64,65,66,67,68,69,70,71,72,73,74,75,76,77,78,79,80,81,82]];
 
 /**
  * @enum {number}
@@ -8950,7 +8950,8 @@ proto.api.v1.Step.ConfigCase = {
   OPENAI_V2: 78,
   LAKEBASE: 79,
   JAVASCRIPTSDKAPI: 80,
-  SNOWFLAKEPOSTGRES: 81
+  SNOWFLAKEPOSTGRES: 81,
+  SNOWFLAKECORTEX: 82
 };
 
 /**
@@ -9071,7 +9072,8 @@ proto.api.v1.Step.toObject = function(includeInstance, msg) {
     openaiV2: (f = msg.getOpenaiV2()) && plugins_restapiintegration_v1_plugin_pb.Plugin.toObject(includeInstance, f),
     lakebase: (f = msg.getLakebase()) && plugins_lakebase_v1_plugin_pb.Plugin.toObject(includeInstance, f),
     javascriptsdkapi: (f = msg.getJavascriptsdkapi()) && plugins_javascriptsdkapi_v1_plugin_pb.Plugin.toObject(includeInstance, f),
-    snowflakepostgres: (f = msg.getSnowflakepostgres()) && plugins_snowflakepostgres_v1_plugin_pb.Plugin.toObject(includeInstance, f)
+    snowflakepostgres: (f = msg.getSnowflakepostgres()) && plugins_snowflakepostgres_v1_plugin_pb.Plugin.toObject(includeInstance, f),
+    snowflakecortex: (f = msg.getSnowflakecortex()) && plugins_restapiintegration_v1_plugin_pb.Plugin.toObject(includeInstance, f)
   };
 
   if (includeInstance) {
@@ -9511,6 +9513,11 @@ proto.api.v1.Step.deserializeBinaryFromReader = function(msg, reader) {
       var value = new plugins_snowflakepostgres_v1_plugin_pb.Plugin;
       reader.readMessage(value,plugins_snowflakepostgres_v1_plugin_pb.Plugin.deserializeBinaryFromReader);
       msg.setSnowflakepostgres(value);
+      break;
+    case 82:
+      var value = new plugins_restapiintegration_v1_plugin_pb.Plugin;
+      reader.readMessage(value,plugins_restapiintegration_v1_plugin_pb.Plugin.deserializeBinaryFromReader);
+      msg.setSnowflakecortex(value);
       break;
     default:
       reader.skipField();
@@ -10186,6 +10193,14 @@ proto.api.v1.Step.serializeBinaryToWriter = function(message, writer) {
       81,
       f,
       plugins_snowflakepostgres_v1_plugin_pb.Plugin.serializeBinaryToWriter
+    );
+  }
+  f = message.getSnowflakecortex();
+  if (f != null) {
+    writer.writeMessage(
+      82,
+      f,
+      plugins_restapiintegration_v1_plugin_pb.Plugin.serializeBinaryToWriter
     );
   }
 };
@@ -13166,6 +13181,43 @@ proto.api.v1.Step.prototype.clearSnowflakepostgres = function() {
  */
 proto.api.v1.Step.prototype.hasSnowflakepostgres = function() {
   return jspb.Message.getField(this, 81) != null;
+};
+
+
+/**
+ * optional plugins.restapiintegration.v1.Plugin snowflakecortex = 82;
+ * @return {?proto.plugins.restapiintegration.v1.Plugin}
+ */
+proto.api.v1.Step.prototype.getSnowflakecortex = function() {
+  return /** @type{?proto.plugins.restapiintegration.v1.Plugin} */ (
+    jspb.Message.getWrapperField(this, plugins_restapiintegration_v1_plugin_pb.Plugin, 82));
+};
+
+
+/**
+ * @param {?proto.plugins.restapiintegration.v1.Plugin|undefined} value
+ * @return {!proto.api.v1.Step} returns this
+*/
+proto.api.v1.Step.prototype.setSnowflakecortex = function(value) {
+  return jspb.Message.setOneofWrapperField(this, 82, proto.api.v1.Step.oneofGroups_[0], value);
+};
+
+
+/**
+ * Clears the message field making it undefined.
+ * @return {!proto.api.v1.Step} returns this
+ */
+proto.api.v1.Step.prototype.clearSnowflakecortex = function() {
+  return this.setSnowflakecortex(undefined);
+};
+
+
+/**
+ * Returns whether this field is set.
+ * @return {boolean}
+ */
+proto.api.v1.Step.prototype.hasSnowflakecortex = function() {
+  return jspb.Message.getField(this, 82) != null;
 };